backend/requirements.txt
# Web framework
fastapi==0.111.1
uvicorn[standard]==0.23.2

# Data validation
pydantic==2.7.1

# Neo4j driver
neo4j>=5.11.0

# Supabase / PostgreSQL
supabase==1.0.1
postgrest>=0.10.6,<0.11.0
httpx>=0.23.0,<0.24.0

# Environment variables
python-dotenv==1.0.0

# HTTP requests
requests==2.32.3

# Typing helpers
typing-extensions>=4.8.0


# --- Added for extended features ---
# Cross-Origin Resource Sharing (for React ↔ FastAPI communication)
fastapi-cors==0.0.6

# Async database support (if you later enable background fetching)
asyncpg==0.29.0

# JSON and utility handling
orjson==3.10.7

# For serving static files (optional, for integration)
aiofiles==23.2.1
